+++ src/constants/P5WrapperClassName.ts
@@ -1,0 +1,1 @@
+export const P5WrapperClassName = "react-p5-wrapper";
