+++ src/contracts/P5CanvasInstance.ts
@@ -1,0 +1,7 @@
+import type p5 from "p5";
+
+import type { SketchProps } from "./SketchProps";
+
+export type P5CanvasInstance<Props extends SketchProps = SketchProps> = p5 & {
+  updateWithProps?: (props: Props) => void;
+};
