+++ src/contracts/P5WrapperProps.ts
@@ -1,0 +1,6 @@
+import type { InputProps } from "./InputProps";
+import type { SketchProps } from "./SketchProps";
+import type { WithChildren } from "./WithChildren";
+
+export type P5WrapperProps<Props extends SketchProps = SketchProps> =
+  WithChildren<InputProps<Props>>;
