+++ src/contracts/SketchProps.ts
@@ -1,0 +1,1 @@
+export type SketchProps = { [key: string]: unknown };
