+++ src/contracts/WithChildren.ts
@@ -1,0 +1,3 @@
+import type { ReactNode } from "react";
+
+export type WithChildren<T = unknown> = T & { children?: ReactNode };
