+++ src/contracts/Wrapper.ts
@@ -1,0 +1,1 @@
+export type Wrapper = HTMLDivElement;
